new folders created. files have been split and every function is in their own file
