+++ controllers/contacts/showContacts.js
@@ -1,0 +1,13 @@
+import { getContactById } from "../../models/contacts";
+
+async function contactById(req, res, next) {
+  const { id } = req.params;
+  const contact = await getContactById(id);
+  if (contact) {
+    res.status(200).json(contact);
+  } else {
+    res.status(404).json({ message: "Not found" });
+  }
+}
+
+export { contactById };
