+++ routes/api/contacts.js
@@ -2,68 +2,26 @@ import express from "express";
 import Joi from "joi";
 const router = express.Router();
 
-import {
-  listContacts,
-  getContactById,
-  addContact,
-  removeContact,
-  updateContact,
-} from "../../models/contacts.js";
+import { indexContacts } from "../../controllers/indexContacts.js";
+import { updateContacts } from "../../controllers/contacts/updateContacts.js";
+import { removeContacts } from "../../controllers/deleteContacts.js";
+import { addNewContact } from "../../controllers/contacts/createContacts.js";
+import { contactById } from "../../controllers/contacts/showContacts.js";
 
-const schema = Joi.object({
+export const schema = Joi.object({
   name: Joi.string().required(),
   email: Joi.string().email().required(),
   phone: Joi.string().required(),
 });
 
-router.get("/", async (req, res, next) => {
-  const contacts = await listContacts();
-  res.status(200).json(contacts);
-});
+router.get("/", indexContacts);
 
-router.get("/:contactId", async (req, res, next) => {
-  const { id } = req.params;
-  const contact = await getContactById(id);
-  if (contact) {
-    res.status(200).json(contact);
-  } else {
-    res.status(404).json({ message: "Not found" });
-  }
-});
+router.get("/:contactId", contactById);
 
-router.post("/", async (req, res, next) => {
-  const result = schema.validate(req.body);
-  if (result.error) {
-    res.status(400).json({ message: result.error.details[0].message });
-  } else {
-    const newContact = await addContact(req.body);
-    res.status(201).json(newContact);
-  }
-});
+router.post("/", addNewContact);
 
-router.delete("/:contactId", async (req, res, next) => {
-  const { id } = req.params;
-  const result = await removeContact(id);
-  if (result) {
-    res.status(200).json({ message: "contact deleted" });
-  } else {
-    res.status(404).json({ message: "Not found" });
-  }
-});
+router.delete("/:contactId", removeContacts);
 
-router.put("/:contactId", async (req, res, next) => {
-  const { id } = req.params;
-  const contactDetails = schema.validate(req.body);
-  if (contactDetails.error) {
-    res.status(400).json({ message: contactDetails.error.details[0].message });
-  } else {
-    const updatedContact = await updateContact(id, { contactDetails });
-    if (updatedContact) {
-      res.status(200).json(updatedContact);
-    } else {
-      res.status(404).json({ message: "Not found" });
-    }
-  }
-});
+router.put("/:contactId", updateContacts);
 
 export { router };
